Issue #SC-211 feat: Created a new actor and backgroundactor for the data security and removed from bulk uopload
